refactor(rln): separate keygen, proof handling, slashing logic, and witness into protocol folder
